tests/wpt/web-platform-tests/html/semantics/interactive-elements/the-popup-element/popup-light-dismiss.tentative.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Popup light dismiss behavior</title>
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<button id=b1 onclick='p1.show()'>Popup 1</button>
<span id=outside>Outside all popups</span>
<popup id=p1 anchor=b1>
  <span id=inside1>Inside popup 1</span>
  <button id=b2 onclick='p2.show()'>Popup 2</button>
</popup>
<popup id=p2 anchor=b2>
  <span id=inside2>Inside popup 2</span>
</popup>

<button id=b3 popup=p3>Popup 3 - button 3
  <popup id=p4>Inside popup 4</popup>
</button>
<popup id=p3>Inside popup 3</popup>
<button id=b4 popup=p3>Popup 3 - button 4
  <popup id=p5>Inside popup 5</popup>
</button>

<popup id=p6>Inside popup 6
  <div style="height:2000px;background:lightgreen"></div>
  Bottom of popup6
</popup>
<button popup=p6>Popup 6</button>

<style>
  #p1 { top:50px; }
  #p2 { top:50px; left:250px; }
  #p3 { top:150px;}
  #p4 { top:150px; left:250px;}
  #p5 { top:250px;}
  #p6 {
    top: 50px;
    left:150px;
    width: 300px;
    height: 300px;
    overflow-y: scroll;
  }
  popup { border: 5px solid red; }
</style>


<script>
  function clickOn(element) {
    const actions = new test_driver.Actions();
    return actions.pointerMove(0, 0, {origin: element})
      .pointerDown({button: actions.ButtonType.LEFT})
      .pointerUp({button: actions.ButtonType.LEFT})
      .send();
  }

  const popup1 = document.querySelector('#p1');
  const button1 = document.querySelector('#b1');
  const popup2 = document.querySelector('#p2');
  const outside = document.querySelector('#outside');
  const inside1 = document.querySelector('#inside1');
  const inside2 = document.querySelector('#inside2');
  const button3 = document.querySelector('#b3');
  const button4 = document.querySelector('#b4');
  const popup3 = document.querySelector('#p3');
  const popup4 = document.querySelector('#p4');
  const popup5 = document.querySelector('#p5');
  const popup6 = document.querySelector('#p6');

  (async function() {
    setup({ explicit_done: true });

    let popup1HideCount = 0;
    popup1.addEventListener('hide',(e) => {
      ++popup1HideCount;
      e.preventDefault(); // 'hide' should not be cancellable.
    });
    let popup2HideCount = 0;
    popup2.addEventListener('hide',(e) => {
      ++popup2HideCount;
      e.preventDefault(); // 'hide' should not be cancellable.
    });

    assert_false(popup1.open);
    popup1.show();
    let p1HideCount = popup1HideCount;
    assert_true(popup1.open);
    await clickOn(outside);
    test(t => {
      assert_false(popup1.open);
      assert_equals(popup1HideCount,p1HideCount+1);
    },'Clicking outside a popup will dismiss the popup');

    assert_false(popup1.open);
    popup1.show();
    p1HideCount = popup1HideCount;
    await clickOn(inside1);
    test(t => {
      assert_true(popup1.open);
      assert_equals(popup1HideCount,p1HideCount);
    },'Clicking inside a popup does not close that popup');

    popup2.show();
    p1HideCount = popup1HideCount;
    let p2HideCount = popup2HideCount;
    await clickOn(inside2);
    test(t => {
      assert_true(popup1.open);
      assert_true(popup2.open);
      assert_equals(popup1HideCount,p1HideCount,'popup1');
      assert_equals(popup2HideCount,p2HideCount,'popup2');
    },'Clicking inside a child popup shouldn\'t close either popup');

    popup2.show();
    p1HideCount = popup1HideCount;
    p2HideCount = popup2HideCount;
    await clickOn(inside1);
    test(t => {
      assert_true(popup1.open);
      assert_equals(popup1HideCount,p1HideCount);
      assert_false(popup2.open);
      assert_equals(popup2HideCount,p2HideCount+1);
    },'Clicking inside a parent popup should close child popup');

    assert_true(popup1.open);
    p1HideCount = popup1HideCount;
    await clickOn(button1);
    test(t => {
      assert_true(popup1.open,'popup1 not open');
      assert_equals(popup1HideCount,p1HideCount);
      popup1.hide(); // Cleanup
      assert_false(popup1.open);
    },'Clicking on anchor element shouldn\'t close its popup');

    await clickOn(button3);
    test(t => {
      assert_true(popup3.open,'invoking element should open popup');
      popup4.show();
      assert_true(popup4.open);
      assert_true(popup3.open);
    },'An invoking element should be part of the ancestor chain');

    await clickOn(button3);
    assert_true(popup3.open);
    assert_false(popup4.open);
    assert_false(popup5.open);
    popup5.show();
    test(t => {
      assert_true(popup5.open);
      assert_false(popup3.open);
    },'An invoking element that was not used to invoke the popup should NOT be part of the ancestor chain');

    popup1.show();
    popup2.show(); // Popup1 is an ancestral element for popup2.
    assert_true(popup1.open);
    assert_true(popup2.open);
    const drag_actions = new test_driver.Actions();
    // Drag *from* popup2 *to* popup1 (its ancestor).
    await drag_actions.pointerMove(0,0,{origin: popup2})
      .pointerDown({button: drag_actions.ButtonType.LEFT})
      .pointerMove(0,0,{origin: popup1})
      .pointerUp({button: drag_actions.ButtonType.LEFT})
      .send();
    test(t => {
      assert_true(popup1.open,'popup1 should be open');
      assert_true(popup2.open,'popup1 should be open');
      popup1.hide();
      assert_false(popup2.open);
    },'Dragging from an open popup outside an open popup should leave the popup open');

    popup6.show();
    assert_equals(popup6.scrollTop,0,'popup6 should start non-scrolled');
    await new test_driver.Actions()
       .scroll(0, 0, 0, 50, {origin: popup6})
       .send();
    test(t => {
      assert_true(popup6.open,'popup6 should stay open');
      assert_equals(popup6.scrollTop,50,'popup6 should be scrolled');
      popup6.hide();
    },'Scrolling within a popup should not close the popup');

    done();
  })();
</script>
